HOCS-6012: New JPA config.

+++ src/main/java/uk/gov/digital/ho/hocs/cms/domain/repository/DocumentsRepository.java
@@ -8,7 +8,7 @@ import uk.gov.digital.ho.hocs.cms.domain.DocumentExtractRecord;
 import java.math.BigDecimal;
 
 @Repository
-public interface DocumentsRepository extends CrudRepository<DocumentExtractRecord, BigDecimal> {
+public interface DocumentsRepository extends CrudRepository<DocumentExtractRecord, Long> {
 
     @Query(value = "SELECT count(*) FROM documents d WHERE d.case_id = ?1 AND d.document_extracted = 'false'", nativeQuery = true)
     BigDecimal findFailedDocumentsForCase(BigDecimal caseId);
